refactor(state): split tests.rs monolith into per-concern files
tests.rs grew to 4714 LOC, 117 #[test] fns, single file. Adding new
test = scroll monolith. Split by concern matching #[path] pattern from
crates/client/src/lib.rs:49-55:

  tests/dag.rs          — 17 tests (insert, equivocation, topo sort, ManagedDag, pending buffer)
  tests/materialize.rs  — 54 tests (channels, messages, profiles, mute, pin, ephemeral, idempotency)
  tests/permissions.rs  — 31 tests (grant/revoke, check_permission, kick, member edits, set_permission round-trip)
  tests/stress.rs       — 7 tests (1000 events, 100 authors, sort perf, governance scale)
  tests/sync.rs         — 2 tests (joining peer + grant batches)
  tests/voting.rs       — 6 tests (proposal/vote ordering, multi-admin kick majority)

All 117 tests preserved verbatim — only relocated. Test count before/after:
227 -> 227 (cargo test -p willow-state).

Sixth bucket (stress.rs) added since the issue's five suggested buckets
don't cleanly absorb the scale tests; CLAUDE.md guidance says add a
sixth file rather than force-fit. Common helpers (genesis_kind, test_dag,
do_emit) duplicated per-file with #![allow(unused_imports, dead_code)]
so each bucket compiles standalone — every bucket only references a
subset, but factoring them into a shared helper module would have
required a load-bearing #[path]-glued helpers.rs that the issue's
mechanical-split scope didn't ask for.

Refs #263
